Title: 최대 숫자 만들기, Time: 811ms, Memory: 176MB, Status: Memory Limit Exceed - CodeTree

+++ 241118/최대 숫자 만들기/make-biggest-num.java
@@ -25,9 +25,9 @@ public class Main {
 			public int compare(String o1, String o2) {
 				String str1 = o1+o2;
 				String str2 = o2+o1;
-				int n1 = Integer.valueOf(str1);
-				int n2 = Integer.valueOf(str2);
-				return Integer.compare(n2, n1);		
+				Long n1 = Long.valueOf(str1);
+				Long n2 = Long.valueOf(str2);
+				return Long.compare(n2, n1);		
 			}
 			
 		});
